project restructured accroding it's new name
