+++ setup.py
@@ -4,7 +4,7 @@ import sys
 from setuptools import setup
 from setuptools.command.test import test as TestCommand
 
-__version__ = '1.2.1'
+__version__ = '1.0.0'
 
 here = os.path.abspath(os.path.dirname(__file__))
 README = open(os.path.join(here, 'README.rst')).read()
@@ -15,7 +15,7 @@ class PyTest(TestCommand):
 
     def initialize_options(self):
         TestCommand.initialize_options(self)
-        self.pytest_args = ['src/test_django_context_logging.py']
+        self.pytest_args = ['src/tests']
 
     def finalize_options(self):
         TestCommand.finalize_options(self)
@@ -28,9 +28,9 @@ class PyTest(TestCommand):
         sys.exit(errno)
 
 setup(
-    name='django-context-logging',
+    name='pylogctx',
     version=__version__,
-    description='Library for adding request context to log records',
+    description='Library for adding context to log records',
     long_description=README,
     classifiers=[
         'Intended Audience :: Developers',
@@ -44,17 +44,15 @@ setup(
     keywords='django context logging',
     author="Lev Orekhov",
     author_email="lev.orekhov@gmail.com",
-    url="https://github.com/lorehov/django_context_logging",
+    url="https://github.com/novafloss/pylogctx",
     license="BSD",
     package_dir={'': 'src'},
-    py_modules=['django_context_logging'],
     cmdclass={'test': PyTest},
-    test_suite="test_django_context_logging",
     include_package_data=True,
     zip_safe=False,
     tests_require=[
         'pytest>=2.9.0',
-        'mock>=2.0.0'
-    ],
-    install_requires=["django>=1.4.3"]
+        'mock>=2.0.0',
+        'django>=1.4.3'
+    ]
 )
