+++ src/tests/ext/test_django.py
@@ -1,16 +1,15 @@
 import copy
 from logging import LogRecord
-import sys
-import threading
 
-import pytest
 from django.core.exceptions import ImproperlyConfigured
 from mock import patch, call
+import pytest
 
-from django_context_logging import (
-    ExtractRequestContextMiddleware, AddContextFormatter, AddContextFilter,
-    context as log_context
+from pylogctx.context import context as log_context
+from pylogctx.ext.django import (
+    ExtractRequestContextMiddleware, AddContextFormatter, AddContextFilter
 )
+from tests.common_fixture import context
 
 
 @pytest.fixture()
@@ -21,16 +20,6 @@ def request():
 @pytest.fixture()
 def record():
     return LogRecord("foo", "INFO", "foo", 10, "waagh :(", (), None)
-
-
-@pytest.yield_fixture
-def context():
-    log_context.update(rid=42)
-    yield None
-    try:
-        del log_context._stack
-    except AttributeError:
-        pass
 
 
 def _extractor(r):
@@ -46,15 +35,15 @@ def test_extract_request_context_middleware_no_extractor(request):
         ExtractRequestContextMiddleware().process_request(request)
 
 
-@patch('django_context_logging.settings',
+@patch('pylogctx.ext.django.settings',
        DJANGO_CONTEXT_LOGGING_EXTRACTOR=_failing_extractor)
 def test_extract_request_context_middleware_extraction_failed(request):
-    with patch('django_context_logging._log') as m:
+    with patch('pylogctx.ext.django._log') as m:
         ExtractRequestContextMiddleware().process_request(request)
         assert m.method_calls == [call.exception()]
 
 
-@patch('django_context_logging.settings',
+@patch('pylogctx.ext.django.settings',
        DJANGO_CONTEXT_LOGGING_EXTRACTOR=_extractor)
 def test_extract_request_context_middleware_context_extracted(request, context):  # noqa
     ExtractRequestContextMiddleware().process_request(request)
@@ -102,66 +91,3 @@ def test_add_context_filter_with_context(record, context):
     assert original_record.__dict__ == record.__dict__
 
 
-def test_update_clear_remove(context):
-    log_context.update(myField='toto', myOtherField='titi')
-    fields = log_context.as_dict()
-    assert 'myField' in fields
-    assert 'myOtherField' in fields
-
-    log_context.remove('myOtherField')
-    fields = log_context.as_dict()
-    assert 'myField' in fields
-    assert 'myOtherField' not in fields
-
-    log_context.clear()
-    fields = log_context.as_dict()
-    assert 'myField' not in fields
-
-
-def test_context_manager(context):
-    # Check two level stack in log context
-    with log_context(myField='toto'):
-        fields = log_context.as_dict()
-        assert 'myField' in fields
-
-        with log_context(myOtherField='toto'):
-            fields = log_context.as_dict()
-            assert 'myOtherField' in fields
-            assert 'myField' in fields
-
-        fields = log_context.as_dict()
-        assert 'myField' in fields
-        assert ('myOtherField', fields)
-
-    fields = log_context.as_dict()
-    assert ('myField', fields)
-
-
-def test_multi_thread(context):
-    # Create a simple child thread updating in log context, but NOT clearing
-    # context.
-    class Child(threading.Thread):
-        def __init__(self, shm):
-            super(Child, self).__init__()
-            self.shm = shm
-
-        def run(self):
-            try:
-                log_context.update(childField='titi')
-                fields = log_context.as_dict()
-                assert 'childField' in fields
-            except (AssertionError, Exception):
-                self.case.shm['thread_exc_info'] = sys.exc_info()
-
-    # Run it
-    shm = dict(thread_exc_info=None)
-    child = Child(shm)
-    child.start()
-    child.join()
-
-    # Ensure child success
-    assert shm['thread_exc_info'] is None
-
-    # Check caller context is safe :-)
-    fields = log_context.as_dict()
-    assert 'childField'not in fields
